Fixup string chunk instead of the whole header, small cleanups

+++ tempesta_fw/http_msg.h
@@ -55,7 +55,7 @@ void tfw_http_msg_iter_init(TfwHttpMsg *hm, TfwMsgIter *it);
 int tfw_http_msg_write(TfwHttpMsg *hm, TfwMsgIter *it, const TfwStr *data);
 
 void tfw_http_msg_hdr_open(TfwHttpMsg *hm, unsigned char *hdr_start);
-void tfw_http_msg_hdr_chunk_fixup(TfwHttpMsg *hm, char *data, long len);
+void tfw_http_msg_hdr_chunk_fixup(TfwHttpMsg *hm, char *data, int len);
 int tfw_http_msg_hdr_close(TfwHttpMsg *hm, int id);
 int tfw_http_msg_grow_hdr_tbl(TfwHttpMsg *hm);
